requirements.txt
altair==4.2.0
astor==0.8.1
attrs==22.1.0
base58==2.1.1
blinker==1.5
boto3==1.26.27
botocore==1.29.27
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
cinemagoer==2022.12.4
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
enum-compat==0.0.3
et-xmlfile==1.1.0
Flask==2.2.2
gitdb==4.0.10
GitPython==3.1.29
greenlet==2.0.1
idna==3.4
IMDbPY==2022.7.9
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.1
numpy==1.23.5
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.45
streamlit==1.16.0
sudo==1.0.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.0
Werkzeug==2.2.2
zipp==3.11.0
